refactor(ui): laxy match config loading, copy tooltip scoreboard test
