+++ fission/src/systems/match_mode/DefaultMatchModeConfigs.ts
@@ -12,10 +12,7 @@ class DefaultMatchModeConfigs {
     private static readonly MANIFEST_LOCATION = `${API_URL}/match_configs/manifest.json`
     private static _configs: MatchModeConfig[] = []
 
-    /** deferring loading so when you import this module, it doesn't fire a request */
-    private static _loading: Promise<void> = new Promise<void>(resolve => {
-        setTimeout(() => resolve(this.load()))
-    })
+    private static _loading: Promise<void> | undefined
 
     static reload(): Promise<void> {
         this._loading = this.load()
@@ -47,6 +44,7 @@ class DefaultMatchModeConfigs {
     }
 
     public static async getConfigs(): Promise<MatchModeConfig[]> {
+        this._loading ??= this.load()
         await this._loading
         return this._configs
     }
